+++ packages/rocketh/src/index.ts
@@ -3,3 +3,5 @@ export * from './executor/types';
 export * from './environment/types';
 export {extendEnvironment, handleSignerProtocol} from './environment';
 export {loadDeployments} from './environment/deployments';
+export * from './environment/utils/artifacts';
+export * from './environment/utils/chains';