+++ packages/rocketh-deploy-proxy/CHANGELOG.md
@@ -1,5 +1,13 @@
 # rocketh-deploy-proxy
 
+## 3.0.4
+
+### Patch Changes
+
+- Updated dependencies
+  - rocketh@0.9.2
+  - rocketh-deploy@3.0.3
+
 ## 3.0.3
 
 ### Patch Changes