+++ packages/rocketh-export/CHANGELOG.md
@@ -1,5 +1,12 @@
 # rocketh-export
 
+## 3.0.2
+
+### Patch Changes
+
+- Updated dependencies
+  - rocketh@0.9.2
+
 ## 3.0.1
 
 ### Patch Changes